+++ src/test/live-coding-playground/src/main/java/playground/AddRAC.java
@@ -1,0 +1,5 @@
+package playground;
+
+public class AddRAC {
+    private final int x
+}