+++ src/test/live-coding-playground/src/main/java/playground/TestSlf4j.java
@@ -3,7 +3,6 @@ package playground;
 
 import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 public class TestSlf4j {
 
     private final String name;
@@ -16,7 +15,7 @@ public class TestSlf4j {
 
 
     public static void main(String[] args) {
-
+        log.debug("");
     }